+++ docs/release-notes.rst
@@ -5,9 +5,11 @@
 Release Notes
 =============
 
-0.1.2
+0.2.0
 -----
 
+* Add progress indication with the option to disable it.
+* Support ``NO_COLOR`` variable in addition to ``NOCOLOR``.
 * [repology/outdated]: fix :bug:`2`, where different packages of the same
   projects crashed the utility.
 * [repology/outdated]: use ``origversion`` if defined to prevent crashes.